ML Demo/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Flask & Web Server
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0  # If deploying with a production server

# Configuration
python-dotenv==1.0.0  # For loading environment variables from .env file

# Azure SDK & Event Hubs
azure-functions==1.17.0
azure-eventhub==5.11.5
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-mgmt-resource
azure-mgmt-storage
azure-mgmt-eventhub
azure-mgmt-web
azure-ai-ml
msrest==0.7.1

# WebSockets
flask-socketio==5.3.6
eventlet==0.33.3  # Needed for async WebSockets
gevent==23.9.1
gevent-websocket==0.10.1

# Data Processing
requests==2.31.0

# Image Processing
pillow
pillow-heif

#run pip install -r requirements.txt to install the dependencies
